% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_ANY_Skills.R
\name{assign_ANY_Skills}
\alias{assign_ANY_Skills}
\title{A SKILL FUNCTION}
\usage{
assign_ANY_Skills(background.Occupation.Skills.ANY = data.frame(),
  background.Social.Skills.ANY = data.frame(), tp.Skills.ANY = data.frame(),
  skills.Increase.Record.DF = data.frame())
}
\arguments{
\item{background.Occupation.Skills.ANY}{DATAFRAME OF THE SKILLS ASSIGNED BY THE PLAYER FROM PARENTS' OCCUPATION. Defaults to empty data frame.}

\item{background.Social.Skills.ANY}{DATAFRAME OF THE SKILLS ASSIGNED BY THE PLAYER FROM PARENTS' SOCIETY AND PLACE WITHIN IT. Defaults to empty data frame.}

\item{tp.Skills.ANY}{DATAFRAME OF THE SKILLS ASSIGNED BY THE PLAYER FROM THE TRAINING PACKAGE. Defaults to empty data frame.}

\item{skills.Increase.Record.DF}{DATAFRAME OF SKILLS WHICH CAN BE INCREASED}
}
\description{
FUNCTION WHICH ASSIGNS SKILLS THAT COME FROM "ANY" OR "UP TO"
}
\examples{
assign_ANY_Skills()
}
\concept{
skill increases
}
