% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_Leftover_DP.R
\name{assign_Leftover_DP}
\alias{assign_Leftover_DP}
\title{A SKILL FUNCTION}
\usage{
assign_Leftover_DP(char.Class = NULL, dev_Points.Level = NULL,
  dev_Points.Spent.Level = NULL, skills.Each.Level = NULL,
  background_Options.Benefits = NULL)
}
\arguments{
\item{char.Class}{STRING CONTAINING CHARACTER'S PROFESSION}

\item{dev_Points.Level}{DEVELPOMENT POINTS FOR EACH LEVEL}

\item{dev_Points.Spent.Level}{DEVELPOMENT POINTS SPENT EACH LEVEL (MUST BE LESS THAN ABOVE)}

\item{skills.Each.Level}{SKILLS INCREASED EACH LEVEL}

\item{background_Options.Benefits}{LIST CONTAINING ALL BACKGROUND OPTION BENEFITS A PLAYER HAS TAKEN}
}
\description{
ASSIGNS LEFTOVER DEVELOPMENT POINTS FROM EACH LEVEL
}
\examples{
assign_Leftover_DP()
}
\concept{
skill increases
}
