% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{background_Options}
\alias{background_Options}
\title{Background Options}
\format{A data frame with 16 rows and 2 variables}
\usage{
background_Options
}
\description{
A dataset containing all background options
}
\details{
\itemize{
  item Number. Placeholder row
  item Option. Name of each option
}
}
\keyword{datasets}
