% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_Options_Assignment_NO_ROLLING.R
\name{background_Options_Assignment_NO_ROLLING}
\alias{background_Options_Assignment_NO_ROLLING}
\title{A BACKGROUND OPTION FUNCTION}
\usage{

  background_Options_Assignment_NO_ROLLING(background_Options.Benefits.ROLLING = -1,
  background_Options.Benefits.NO.ROLLING = -1,
  background.Option.Points = -1, stats.Tmp.Pot = -1, stats = -1,
  background_Options = -1, skills.Primary.DF = -1,
  skills.Secondary.DF = -1, skills.Increase.Record.DF = -1,
  background_Special_Tables_List = -1, flaws_Table = -1,
  background.Skills.Increase.Record.DF = -1)
}
\arguments{
\item{background_Options.Benefits.ROLLING}{LIST CONTAINING ALL BACKGROUND OPTION BENEFITS A PLAYER HAS TAKEN WHICH REQUIRED ROLLING}

\item{background_Options.Benefits.NO.ROLLING}{LIST CONTAINING ALL BACKGROUND OPTION BENEFITS A PLAYER HAS TAKEN WHICH DID NOT REQUIRE ROLLING}

\item{background.Option.Points}{NUMERIC VARIABLE CONTAINING THE NUMBER OF BACKGROUND OPTION POINTS A CHARACTER HAS}

\item{stats.Tmp.Pot}{MATRIX OF TEMPORARY AND POTENTIAL STATS}

\item{stats}{VECTOR OF ALL 10 STATS}

\item{background_Options}{DF OF BACKGROUND OPTIONS AVAILABLE}

\item{skills.Primary.DF}{DATAFRAME OF THE PRIMARY SKILLS}

\item{skills.Secondary.DF}{DATAFRAME OF THE SECONDARY SKILLS}

\item{skills.Increase.Record.DF}{DATAFRAME OF SKILLS YOU CAN INCREASE WHICH WILL GENERATE ONES USED TO RECORD ANY INCREASES}

\item{background_Special_Tables_List}{TABLE CONTAINING ALL SPECIAL "ROLL ON" TABLE THE CHARACTER COULD USE}

\item{flaws_Table}{TABLE CONTAINING ALL FLAWS THE CHARACTER COULD TAKE}

\item{background.Skills.Increase.Record.DF}{DATAFRAME OF SKILLS YOU CAN INCREASE WHICH WILL GENERATE ONES USED TO RECORD ANY INCREASES AT LEVEL 0}
}
\description{
LETS THE PLAYER CHOOSE THEIR BACKGROUND OPTIONS AND GENERATES THE BENEFITS
}
\examples{
background_Options_Assignment_NO_ROLLING()
}
\concept{
background options
}
