% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_Options_Assignment_ROLLING.R
\name{background_Options_Assignment_ROLLING}
\alias{background_Options_Assignment_ROLLING}
\title{A BACKGROUND OPTION FUNCTION}
\usage{

  background_Options_Assignment_ROLLING(background_Options.Benefits.ROLLING = -1,
  background.Option.Points = -1, background_Options = -1,
  background_Special_Tables_List = -1, flaws_Table = -1)
}
\arguments{
\item{background_Options.Benefits.ROLLING}{LIST CONTAINING ALL BACKGROUND OPTION BENEFITS A PLAYER HAS TAKEN WHICH REQUIRED ROLLING}

\item{background.Option.Points}{NUMERIC VARIABLE CONTAINING THE NUMBER OF BACKGROUND OPTION POINTS A CHARACTER HAS}

\item{background_Options}{DF OF BACKGROUND OPTIONS AVAILABLE}

\item{background_Special_Tables_List}{TABLE CONTAINING ALL SPECIAL "ROLL ON" TABLE THE CHARACTER COULD USE}

\item{flaws_Table}{TABLE CONTAINING ALL FLAWS THE CHARACTER COULD TAKE}
}
\description{
LETS THE PLAYER CHOOSE THEIR BACKGROUND OPTIONS THAT REQUIRE ROLLING AND GENERATES THE BENEFITS
}
\examples{
background_Options_Assignment_ROLLING()
}
\concept{
background options
}
