% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_Options_Choice_NO_ROLLING.R
\name{background_Options_Choice_NO_ROLLING}
\alias{background_Options_Choice_NO_ROLLING}
\title{A BACKGROUND OPTIONS FUNCTION}
\usage{
background_Options_Choice_NO_ROLLING(background_Options = -1,
  skills.Secondary.DF = -1, skills.Primary.DF = -1,
  background_Special_Tables_List = -1, flaws_Table = -1, stat.Tmp = -1,
  stat.Pot = -1, skills.Increase.Record.DF = -1,
  background_Options.Benefits = -1)
}
\arguments{
\item{background_Options}{WHICH BACKGROUND OPTIONS ARE AVAILABLE}

\item{skills.Secondary.DF}{DATAFRAME OF THE SECONDARY SKILLS}

\item{skills.Primary.DF}{DATAFRAME OF THE PRIMARY SKILLS}

\item{background_Special_Tables_List}{TABLE CONTAINING ALL SPECIAL "ROLL ON" TABLE THE CHARACTER COULD USE}

\item{flaws_Table}{TABLE CONTAINING ALL FLAWS THE CHARACTER COULD TAKE}

\item{stat.Tmp}{CHARACTER'S TEMPORARY STATS}

\item{stat.Pot}{CHARACTER'S POTENTIAL STATS}

\item{skills.Increase.Record.DF}{DATAFRAME OF SKILLS YOU CAN INCREASE WHICH WILL GENERATE ONES USED TO RECORD ANY INCREASES}

\item{background_Options.Benefits}{LIST CONTAINING ALL BACKGROUND OPTION BENEFITS A PLAYER HAS TAKEN}
}
\description{
FUNCTION TO DEAL WITH BACKGROUND OPTION CHOICES
}
\examples{
background_Options_Choice_NO_ROLLING()
}
\concept{
skill increases
}
