% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_Options_Choice_ROLLING.R
\name{background_Options_Choice_ROLLING}
\alias{background_Options_Choice_ROLLING}
\title{A BACKGROUND OPTIONS FUNCTION}
\usage{
background_Options_Choice_ROLLING(background_Options = -1,
  background_Special_Tables_List = -1, flaws_Table = -1,
  background_Options.Benefits = -1)
}
\arguments{
\item{background_Options}{WHICH BACKGROUND OPTIONS ARE AVAILABLE}

\item{background_Special_Tables_List}{TABLE CONTAINING ALL SPECIAL "ROLL ON" TABLE THE CHARACTER COULD USE}

\item{flaws_Table}{TABLE CONTAINING ALL FLAWS THE CHARACTER COULD TAKE}

\item{background_Options.Benefits}{LIST CONTAINING ALL BACKGROUND OPTION BENEFITS A PLAYER HAS TAKEN}
}
\description{
FUNCTION TO DEAL WITH INDIVIDUAL BACKGROUND OPTION CHOICES THAT REQUIRE ROLLING
}
\examples{
background_Options_Choice_NO_ROLLING()
}
\concept{
background options
}
