% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background_Options_Choices_Combine.R
\name{background_Options_Choices_Combine}
\alias{background_Options_Choices_Combine}
\title{A BACKGROUND OPTIONS FUNCTION FUNCTION}
\usage{
background_Options_Choices_Combine(background_Options.Benefits.ROLLING = -1,
  background_Options.Benefits.NO.ROLLING = -1)
}
\arguments{
\item{background_Options.Benefits.ROLLING}{OPTIONS WHICH REQUIRED ROLLING CHOSEN WITH BACKGROUND OPTION POINTS}

\item{background_Options.Benefits.NO.ROLLING}{OPTIONS WHICH DID NOT REQUIR ROLLING CHOSEN WITH BACKGROUND OPTION POINTS}
}
\description{
FUNCTION TO COMBINE BACKGROUND BENEFITS FROM ROLLING AND CHOOSING
}
\examples{
background_Options_Choices_Combine()
}
\concept{
background options
}
