% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{background_Special_Abilities}
\alias{background_Special_Abilities}
\title{Special Abilities Table}
\format{A data frame with 23 rows and 3 variables}
\usage{
background_Special_Abilities
}
\description{
A dataset containing the chances of getting a certain ability from the special abilities table during background creation
}
\details{
\itemize{
  item Roll. Upper bound of roll to be in this row
  item Ability. What is their special ability
  item Description. Description of what this special ability means
}
}
\keyword{datasets}
