% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{background_Special_Curses}
\alias{background_Special_Curses}
\title{Special Curse Table}
\format{A data frame with 15 rows and 2 variables}
\usage{
background_Special_Curses
}
\description{
A dataset containing the chances of getting a certain curse from the special curses table during background creation
}
\details{
\itemize{
  item Roll. Upper bound of roll to be in this row
  item Family.Tie. Resulting curse
}
}
\keyword{datasets}
