% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{background_Special_Family_Ties}
\alias{background_Special_Family_Ties}
\title{Special Family Tie Table}
\format{A data frame with 11 rows and 2 variables}
\usage{
background_Special_Family_Ties
}
\description{
A dataset containing the chances of getting a certain family tie from the special family ties table during background creation
}
\details{
\itemize{
  item Roll. Upper bound of roll to be in this row
  item Family.Tie. Resulting family tie
}
}
\keyword{datasets}
