% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{background_Special_Items}
\alias{background_Special_Items}
\title{Special Item Table}
\format{A data frame with 15 rows and 2 variables}
\usage{
background_Special_Items
}
\description{
A dataset containing the chances of getting a certain item from the special items table during background creation
}
\details{
\itemize{
  item Roll. Upper bound of roll to be in this row
  item Item. Resulting item
}
}
\keyword{datasets}
