% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{background_Special_Status}
\alias{background_Special_Status}
\title{Special Status Table}
\format{A data frame with 17 rows and 2 variables}
\usage{
background_Special_Status
}
\description{
A dataset containing the chances of getting a certain status from the special status table during background creation
}
\details{
\itemize{
  item Roll. Upper bound of roll to be in this row
  item Wealth. Resulting status
}
}
\keyword{datasets}
