% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{background_Special_Wealth}
\alias{background_Special_Wealth}
\title{Special Wealth Table}
\format{A data frame with 17 rows and 2 variables}
\usage{
background_Special_Wealth
}
\description{
A dataset containing the chances of getting certain wealth from the special wealth table during background creation
}
\details{
\itemize{
  item Roll. Upper bound of roll to be in this row
  item Wealth. Resulting wealth
}
}
\keyword{datasets}
