% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_Height_And_Weight.R
\name{calculate_Height_And_Weight}
\alias{calculate_Height_And_Weight}
\title{A HEIGHT AND WEIGHT FUNCTION}
\usage{
calculate_Height_And_Weight(char.Species = -1, char.Sex = -1,
  char.Height = -1, char.Weight = -1)
}
\arguments{
\item{char.Species}{STRING CONTAINING CHARACTER'S SPECIES}

\item{char.Sex}{STRING CONTAINING CHARACTER'S SEX}

\item{char.Height}{= -1 NUMERIC VARIABLE CONTAINING CHARACTER'S HEIGHT}

\item{char.Weight}{= -1 NUMERIC VARIABLE CONTAINING CHARACTER'S WEIGHT}
}
\description{
CACULATES HEIGHT, WEIGHT, AND MOVE MOD FOR THE CHARACTER
}
\examples{
calculate_Height_And_Weight()
}
\concept{
character description
}
