% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_All_Stats_Assign.R
\name{char_All_Stats_Assign}
\alias{char_All_Stats_Assign}
\title{A STAT FUNCTION}
\usage{
char_All_Stats_Assign(stats.Tmp = -1, stats.Mat = -1, stats = -1,
  resist.Realm = "", adept.Realm = "", char.Class.FullName = "")
}
\arguments{
\item{stats.Tmp}{CHARACTER'S TEMPORARY STATS}

\item{stats.Mat}{CHARACTER'S CURRENTLY ASSIGNED TEMPORARY STATS}

\item{stats}{VECTOR OF THE 10 STATS}

\item{resist.Realm}{SPELLCASTING REALM CHARACTER IS RESISTANT TO}

\item{adept.Realm}{SPELLCASTING REALM CHARACTER IS ADEPT WITH}

\item{char.Class.FullName}{CHARACTER'S PROFESSION'S FULL NAME}
}
\description{
FUNCTION TO ASSIGN THE STAT ROLLS TO SPECIFIC STATS
}
\examples{
char_All_Stats_Assign()
}
\concept{
skill increases
}
