% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_Basic_Info.R
\name{char_Basic_Info}
\alias{char_Basic_Info}
\title{A CHARACTER BUILDING FUNCTION}
\usage{
char_Basic_Info(char.Species = -1, char.Name = -1, char.Player = -1,
  char.Sex = -1, char.Society = -1, char.Level = -1,
  char_Social_Class = -1, species_Description_DF = -1)
}
\arguments{
\item{char.Species}{SPECIES THE CHARACTER IS (-1 MEANS YET TO BE DECIDED)}

\item{char.Name}{NAME OF THE CHARACTER (-1 MEANS YET TO BE DECIDED)}

\item{char.Player}{NAME OF THE PLAYER (-1 MEANS YET TO BE DECIDED)}

\item{char.Sex}{SEX OF THE CHARACTER (-1 MEANS YET TO BE DECIDED)}

\item{char.Society}{SOCIETY THE CHARACTER GREW UP IN (-1 MEANS YET TO BE DECIDED)}

\item{char.Level}{LEVEL OF THE CHARACTER (-1 MEANS YET TO BE DECIDED)}

\item{char_Social_Class}{DATAFRAME OF POSSIBLE SOCIAL CLASSES THE CHARACTER CAN HAVE}

\item{species_Description_DF}{DATAFRAME OF POSSIBLE SPECIES THE CHARACTER CAN BE}
}
\description{
GETS THE SPECIES, NAME, PLAYER, SEX AND SOCIETY FOR THE CHARACTER
}
\examples{
char_Basic_Info()
}
\concept{
species name player sex society level
}
