% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{char_Clan_Status}
\alias{char_Clan_Status}
\title{Character and Clan Relationship}
\format{A data frame with 6 rows and 3 variables}
\usage{
char_Clan_Status
}
\description{
A dataset containing the chances of the clanhead being viewed in various ways by the clan
}
\details{
\itemize{
  item Roll. Upper bound of roll to be in this row
  item Status. What is their status
  item Description. Description of what this status means
}
}
\keyword{datasets}
