% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{char_Clanhead_Relationship}
\alias{char_Clanhead_Relationship}
\title{Character and Clanhead Relationship}
\format{A data frame with 6 rows and 2 variables}
\usage{
char_Clanhead_Relationship
}
\description{
A dataset containing the chances of the clanhead being related in various ways or unrelated to the character
}
\details{
\itemize{
  item Roll. Upper bound of roll to be in this row
  item Status. What is their status
}
}
\keyword{datasets}
