% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_Class_Decisions_Stat_Assignment.R
\name{char_Class_Decisions_Stat_Assignment}
\alias{char_Class_Decisions_Stat_Assignment}
\title{A CLASS AND STAT FUNCTION}
\usage{
char_Class_Decisions_Stat_Assignment(char.Class = -1, resist.Realm = -1,
  adept.Realm = -1, realms = -1, realms.Stats = -1,
  realms.Stats.Short = -1, classes = -1, stat.Mat = -1, char.Realm = -1)
}
\arguments{
\item{char.Class}{CHARACTER'S CLASS}

\item{resist.Realm}{SPELLCASTING REALM CHARACTER IS RESISTANT TO}

\item{adept.Realm}{SPELLCASTING REALM CHARACTER IS ADEPT WITH}

\item{realms}{STRING CONTAINING THE THREE SPELLCASTING REALMS}

\item{realms.Stats}{STRING CONTAINING THE STATS ASSOCIATED THREE SPELLCASTING REALMS}

\item{realms.Stats.Short}{STRING CONTAINING THE SHORTENED STATS ASSOCIATED THREE SPELLCASTING REALMS}

\item{classes}{STRING CONTAINING ALL PROFESSIONS}

\item{stat.Mat}{MATRIX OF CHARACTER'S STATS}

\item{char.Realm}{CHARACTER'S SPELLCASTING REALM}
}
\description{
FUNCTION TO CHOOSE CLASS, REALM AND ASSIGN STATS
}
\examples{
char_Class_Decisions_Stat_Assignment()
}
\concept{
level character
}
