% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{char_Contacts_Identity}
\alias{char_Contacts_Identity}
\title{Contact Identity}
\format{A data frame with 6 rows and 2 variables}
\usage{
char_Contacts_Identity
}
\description{
A dataset containing the chances of each contact being from a certain social area
}
\details{
\itemize{
  item Roll. Upper bound of roll to be in this row
  item Status. What is their status
}
}
\keyword{datasets}
