% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{char_Contacts_Number}
\alias{char_Contacts_Number}
\title{Contact Relationship Status}
\format{A data frame with 12 rows and 2 variables}
\usage{
char_Contacts_Number
}
\description{
A dataset containing the chances of character being in different states with their contactss
}
\details{
\itemize{
  item Race. Name of each race
  item Divisor. Divide character's starting age by this to work out their number of starting contacts
}
}
\keyword{datasets}
