% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{char_Contacts_Relationship}
\alias{char_Contacts_Relationship}
\title{Contact Relationship Status}
\format{A data frame with 5 rows and 3 variables}
\usage{
char_Contacts_Relationship
}
\description{
A dataset containing the chances of character being in different states with their contactss
}
\details{
\itemize{
  item Roll. Upper bound of roll to be in this row
  item Status. What is their status
  item Loyalty. Base loyalty plus a roll of an n sided die
}
}
\keyword{datasets}
