% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{char_Family_Status}
\alias{char_Family_Status}
\title{Family Status}
\format{A data frame with 6 rows and 2 variables}
\usage{
char_Family_Status
}
\description{
A dataset containing the chances of character being in different family states
}
\details{
\itemize{
  item Roll. Upper bound of roll to be in this row
  item Status. What is their status
}
}
\keyword{datasets}
