% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{char_Height_And_Weight}
\alias{char_Height_And_Weight}
\title{Height and Weight Calcs}
\format{A data frame with 16 rows and 6 variables}
\usage{
char_Height_And_Weight
}
\description{
A dataset containing the numbers needed to caculate height and weight
}
\details{
\itemize{
  item Race. Name of each race
  item Gender. Split into sexes
  item Base.Height. Base height for each race
  item Height.Mod. Height mod for each race
  item Base.Weight. Base weight for each race
  item Weight.Mod. Weight mod for each race
}
}
\keyword{datasets}
