% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{char_Movement_Mods}
\alias{char_Movement_Mods}
\title{Movement Mods from Height}
\format{A data frame with 10 rows and 2 variables}
\usage{
char_Movement_Mods
}
\description{
A dataset containing the mods to base speed for different heights
}
\details{
\itemize{
  item Height. How tall is the character
  item Modification. Mod to base speed
}
}
\keyword{datasets}
