% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{char_No_of_Siblings}
\alias{char_No_of_Siblings}
\title{Number of Siblings}
\format{A data frame with 12 rows and 4 variables}
\usage{
char_No_of_Siblings
}
\description{
A dataset containing the chances having different numbers of siblings depending on race
}
\details{
\itemize{
  item Race. Name of each race
  item Roll_Mod. Number to subtract (meaning you are more likely to have a lower number of siblings)
  item Sibling_Dice. Number of sides on the die which determines how many siblings you have
  item Chance_Alive. Target to roll under on a d100 for them to be alive
}
}
\keyword{datasets}
