% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{char_Orphan_Status}
\alias{char_Orphan_Status}
\title{Orphan Status}
\format{A data frame with 3 rows and 2 variables}
\usage{
char_Orphan_Status
}
\description{
A dataset containing the chances of being fostered, adopted or living alone if you're an orphan
}
\details{
\itemize{
  item Roll. Upper bound of roll to be in this row
  item Status. What is their status
}
}
\keyword{datasets}
