% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{char_Parents_Status}
\alias{char_Parents_Status}
\title{Parent Status}
\format{A data frame with 6 rows and 2 variables}
\usage{
char_Parents_Status
}
\description{
A dataset containing the chances of parents being in different states
}
\details{
\itemize{
  item Roll. Upper bound of roll to be in this row
  item Status. What is their status
}
}
\keyword{datasets}
