% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{char_Profession_Bonus}
\alias{char_Profession_Bonus}
\title{Profession Bonuses}
\format{A data frame with 19 rows and 3 variables}
\usage{
char_Profession_Bonus
}
\description{
A dataset containing the bonuses to skill totals for each profession
}
\details{
\itemize{
  item Profession. Name for each profession 
  item Profession.Bonuses. Skills which have bonuses for each profession
  item Bonuses. Bonuses for each profession
}
}
\keyword{datasets}
