% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_Realm_Relations.R
\name{char_Realm_Relations}
\alias{char_Realm_Relations}
\title{A REALM FUNCTION}
\usage{
char_Realm_Relations(background_Options.Benefits.ROLLING = -1, realms = -1)
}
\arguments{
\item{background_Options.Benefits.ROLLING}{A LIST CONTAINING THE RESULTS OF ALL THE BENEFITS WHICH HAVE TO BE ROLLED FOR, AND ANY FLAWS TAKEN}

\item{realms}{A STRING CONTAINING THE THREE SPELLCASTING REALMS}
}
\description{
CHECKS TO SEE IF CHARACTERS ARE PARTICULARLY RESISTANT TO OR ADEPT WITH ANY REALM
}
\examples{
character_LEVEL()
}
\concept{
realms
}
