% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_Realm_Stat_Assignment.R
\name{char_Realm_Stat_Assignment}
\alias{char_Realm_Stat_Assignment}
\title{A REALM STAT FUNCTION}
\usage{
char_Realm_Stat_Assignment(char.Class = -1, classes = -1)
}
\arguments{
\item{char.Class}{STRING CONTAINING THE CHARACTER'S PROFESSION}

\item{classes}{VECTOR CONTAINING THE POSSIBLE CLASSES}
}
\description{
FUNCTION TO GET THE PLAYER'S REALM STAT
}
\examples{
char_Realm_Stat_Assignment()
}
\concept{
level character
}
