% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{char_Sibling_Rank}
\alias{char_Sibling_Rank}
\title{Social class}
\format{A data frame with 6 rows and 2 variables}
\usage{
char_Sibling_Rank
}
\description{
A dataset containing the chances of being in each social class
}
\details{
\itemize{
  item Roll. Upper bound of roll to be in this row
  item Rank. Which child were you
}
}
\keyword{datasets}
