% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_Single_Stat_Assign.R
\name{char_Single_Stat_Assign}
\alias{char_Single_Stat_Assign}
\title{A STAT FUNCTION}
\usage{
char_Single_Stat_Assign(stat.Tmp = -1, stat.Mat = -1, stats = -1,
  i = -1, resist.Realm = "", adept.Realm = "", char.Class.FullName = "")
}
\arguments{
\item{stat.Tmp}{REMAINING STAT ROLLS TO ASSIGN}

\item{stat.Mat}{STATS ASSIGNED, 0s ELSEWHERE}

\item{stats}{NAME OF ALL STATS}

\item{i}{WHICH STAT IS BEING SET}

\item{resist.Realm}{SPELLCASTING REALM CHARACTER IS RESISTANT TO}

\item{adept.Realm}{SPELLCASTING REALM CHARACTER IS ADEPT WITH}

\item{char.Class.FullName}{CHARACTER'S CLASS'S FULL NAME}
}
\description{
FUNCTION TO ASSIGN ONE OF THE STAT VALUES ROLLED TO ONE OF THE "Co, Ag, SD, etc..."
}
\examples{
char_Single_Stat_Assign()
}
\concept{
level character
}
