% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{char_Skills_Background}
\alias{char_Skills_Background}
\title{Background Skills}
\format{A data frame with 19 rows and 11 variables}
\usage{
char_Skills_Background
}
\description{
A dataset containing the skills from being raised in a certain background
}
\details{
\itemize{
  item Skill. Name of each skills
  item Urban.Man. Skills for being raised by Tribal people
  item Rural.Man. Skills for being raised by Tribal people
  item Noble.Man. Skills for being raised by Tribal people
  item Tribal.Man. Skills for being raised by Tribal people
  item Wood.Elf. Skills for being raised by Wood.Elf people
  item High.Elf. Skills for being raised by High.Elf people
  item Dwarf. Skills for being raised by Dwarf people
  item Hobbit. Skills for being raised by Hobbit people
  item Gnome. Skills for being raised by Gnome people
}
}
\keyword{datasets}
