% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{char_Social_Class}
\alias{char_Social_Class}
\title{Social class}
\format{A data frame with 5 rows and 9 variables}
\usage{
char_Social_Class
}
\description{
A dataset containing the chances of being in each social class
}
\details{
\itemize{
  item Class. Name of each class
  item Tribal. Upper bound on roll to be in each class for Tribal people
  item Viking. Upper bound on roll to be in each class for Viking people
  item Feudal. Upper bound on roll to be in each class for Feudal people
  item Imperial. Upper bound on roll to be in each class for Imperial people
  item Elf. Upper bound on roll to be in each class for Elf people
  item Dwarf. Upper bound on roll to be in each class for Dwarf people
  item Hobbit. Upper bound on roll to be in each class for Hobbit people
  item Gnome. Upper bound on roll to be in each class for Gnome people
}
}
\keyword{datasets}
