% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_Species_Occupation.R
\name{char_Species_Occupation}
\alias{char_Species_Occupation}
\title{AN OCCUPATION FUNCTION}
\usage{
char_Species_Occupation(background.Social_Class = -1, occupation.List = -1,
  background.Occupation = -1)
}
\arguments{
\item{background.Social_Class}{VARIABLE THAT DETERMINES WHERE THEY ARE ON THE SLAVE-NOBLE SPECTRUM.}

\item{occupation.List}{LIST CONTAINING ALL POSSIBLE OCCUPATIONS, WITH A DATAFRAME FOR EACH CLASS INCLUDED.}

\item{background.Occupation}{CONTAINS EITHER THE OCCUPATION OF THE CHARACTER'S PARENTS OR -1}
}
\description{
FUNCTION TO GET THE PLAYER'S OCCUPATION BASED ON THEIR SPECIES/SOCIAL CLASS
}
\examples{
char_Species_Occupation()
}
\concept{
occupation
}
