% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_Species_Raised_Skills_And_Hobby_Points.R
\name{char_Species_Raised_Skills_And_Hobby_Points}
\alias{char_Species_Raised_Skills_And_Hobby_Points}
\title{A SPECIES BORN AND SPECIES RAISED FUNCTION}
\usage{
char_Species_Raised_Skills_And_Hobby_Points(char.Species.Raised = -1,
  char_Skills_Background = -1, char.Society = -1,
  background.Social_Class = -1, urban.Chance = -1)
}
\arguments{
\item{char.Species.Raised}{WHAT SPECIES THEY WERE RAISED AS}

\item{char_Skills_Background}{WHAT SKILLS ARE AVAILABLE FROM DIFFERENT BACKGROUNDS}

\item{char.Society}{WHAT KIND OF SOCIETY ARE THEY FROM?}

\item{background.Social_Class}{WHAT CLASS ARE THEY FROM - GUILDED, UNGUILDED ETC.}

\item{urban.Chance}{WHAT CHANCE DOES THAT JOB HAVE OF BEING URBAN?}
}
\description{
FUNCTION WHICH DETERMINES WHAT SKILL AND HOBBY POINTS THEIR BACKGROUND GIVES THEM
}
\examples{
char_Species_Raised_Skills_And_Hobby_Points()
}
\concept{
level character
}
