% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_Stat_Generate.R
\name{char_Stat_Generate}
\alias{char_Stat_Generate}
\title{A STAT GENERATOR FUNCTION}
\usage{
char_Stat_Generate(resist.Realm = "", adept.Realm = "",
  char.Class.FullName = "Commoner")
}
\arguments{
\item{resist.Realm}{SPELLCASTING REALM CHARACTER IS RESISTANT TO}

\item{adept.Realm}{SPELLCASTING REALM CHARACTER IS ADEPT WITH}

\item{char.Class.FullName}{= "Commoner" STRING CONTAINING THE FULL NAME OF THE CHARACTER'S PROFESSION. DEFAULT TO COMMONER}
}
\description{
FUNCTION TO GENERATE AND ASSIGN ALL STATS THE CHARACTER NEEDS
}
\examples{
char_Stat_Generate()
}
\concept{
output pdf img png character
}
