% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/char_Use.R
\name{char_Use}
\alias{char_Use}
\title{A CHARACTER SHEET GENERATOR FUNCTION}
\usage{
char_Use(EXAMPLE = FALSE)
}
\arguments{
\item{EXAMPLE}{BOOLEAN VARIABLE. IF IT'S AN EXAMPLE, IT RETURNS -1 SO THAT THE PACKAGE CAN BE CHECKED.}
}
\description{
FUNCTION WHICH CREATES, ALTERS OR LEVELS A CHARACTER AND THEN GENERATES A CHARACTER SHEET FOR THEM
}
\examples{
char_Use( EXAMPLE = TRUE )
}
\concept{
character level alter create
}
