% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_ALTER.R
\name{character_ALTER}
\alias{character_ALTER}
\title{A CHARACTER SHEET GENERATOR FUNCTION}
\usage{
character_ALTER(char.Name = -1)
}
\arguments{
\item{char.Name}{NAME OF THE CHARACTER  TO LEVEL UP}
}
\description{
FUNCTION WHICH ALTERS AN ALREADY CREATED CHARACTER AND THEN GENERATES A CHARACTER SHEET
}
\examples{
character_ALTER()
}
\concept{
level character
}
