% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_CREATE.R
\name{character_CREATE}
\alias{character_CREATE}
\title{A CHARACTER SHEET CREATOR FUNCTION}
\usage{
character_CREATE(char.Species = -1, char.Name = -1, char.Player = -1,
  char.Sex = -1, char.Society = -1, char.Class = -1, char.Level = -1,
  stat.Mat = -1, char.Realm = -1, background.No_of_Children = -1,
  background.Sibling_Rank = -1, background.Family_Status = -1,
  background.Environment.Raised = -1, background.Clanhead.Relation = -1,
  background.Clan.Status = -1, background.Social_Class = -1,
  background.Occupation = -1, char.Age = -1, char.Height = -1,
  char.Weight = -1, contact.Info = -1,
  background_Options.Benefits.ROLLING = -1,
  background_Options.Benefits.NO.ROLLING = -1, tp.Choice = -1,
  level_0_Skills.Increase.Record.List = -1,
  level_1_Skills.Increase.Record.List = -1, char.Age.List = -1,
  skills.Each.Level.List = -1, level_Up.Result.List = -1,
  level_Up.Repeated.Result.List = -1, img.Size = "MED", EXAMPLE = FALSE)
}
\arguments{
\item{char.Species}{SPECIES THE CHARACTER IS (-1 MEANS YET TO BE DECIDED)}

\item{char.Name}{NAME OF THE CHARACTER (-1 MEANS YET TO BE DECIDED)}

\item{char.Player}{NAME OF THE PLAYER (-1 MEANS YET TO BE DECIDED)}

\item{char.Sex}{THE SEX OF THE CHARACTER (-1 MEANS YET TO BE DECIDED)}

\item{char.Society}{WHAT KIND OF SOCIETY ARE THEY FROM? (-1 MEANS YET TO BE DECIDED)}

\item{char.Class}{CHARACTER'S CLASS (-1 MEANS YET TO BE DECIDED)}

\item{char.Level}{LEVEL OF THE CHARACTER (-1 MEANS YET TO BE DECIDED)}

\item{stat.Mat}{MATRIX OF TEMPORARY AND POTENTIAL STATS (-1 MEANS YET TO BE DECIDED)}

\item{char.Realm}{CHARACTER'S SPELLCASTING REALM (-1 MEANS YET TO BE DECIDED)}

\item{background.No_of_Children}{DATAFRAME CONTAINING THE WAYS TO CALCULATE THE NUMBER OF CHILDREN FOR EACH SPECIES}

\item{background.Sibling_Rank}{DATAFRAME CONTAINING THE WAYS TO CALCULATE THE NUMBER CHILD THE PLAYER WAS}

\item{background.Family_Status}{DATAFRAME CONTAINING THE DIFFERENT POSSIBLE STATUSES FOR A FAMILY FOR EACH SPECIES, e.g., OFFSPRING, ADOPTED, etc.}

\item{background.Environment.Raised}{DATAFRAME CONTAINING THE DIFFERENT POSSIBLE FAMILY ENVIRONMENTS FOR EACH SPECIES, e.g., BASTARD, STILL TOGETHER, etc.}

\item{background.Clanhead.Relation}{DATAFRAME CONTAINING THE DIFFERENT POSSIBLE RELATIONSHIPS BETWEEN THE CHARACTER AND THE CLANHEAD}

\item{background.Clan.Status}{DATAFRAME CONTAINING THE DIFFERENT POSSIBLE STATUSES THE CHARACTER CAN HAVE IN THE CLAN}

\item{background.Social_Class}{DATAFRAME CONTAINING THE DIFFERENT POSSIBLE SOCIAL CLASSES OF THE FAMILY A CHARACTER IS RAISED IN}

\item{background.Occupation}{DATAFRAME CONTAINING THE DIFFERENT POSSIBLE OCCUPATIONS OF THE FAMILY A CHARACTER IS RAISED IN}

\item{char.Age}{= -1 CHARACTER'S CURRENT AGE}

\item{char.Height}{NUMERIC CHARACTER'S HEIGHT}

\item{char.Weight}{NUMERIC CHARACTER'S WEIGHT}

\item{contact.Info}{= -1 DATAFRAME CONTAINING THE DIFFERENT POSSIBLE OCCUPATIONS OF THE FAMILY A CHARACTER IS RAISED IN}

\item{background_Options.Benefits.ROLLING}{LIST CONTAINING ALL BACKGROUND OPTION BENEFITS A PLAYER HAS TAKEN}

\item{background_Options.Benefits.NO.ROLLING}{LIST CONTAINING ALL BACKGROUND OPTION BENEFITS A PLAYER HAS TAKEN}

\item{tp.Choice}{EITHER A LIST CONTAINING ALL TP RESULTS OR -1}

\item{level_0_Skills.Increase.Record.List}{LIST CONTAINING ALL THE level_0 INFORMATION IF THAT HAS BEEN PREGENERATED}

\item{level_1_Skills.Increase.Record.List}{RESULT OF LEVELING UP AT LEVEL 1}

\item{char.Age.List}{LIST CONTAINING THE CHARACTER'S AGE AT EACH LEVEL}

\item{skills.Each.Level.List}{LIST CONTAINING DATAFRAMES FOR THE SKILLS LEVELED UP EACH LEVEL}

\item{level_Up.Result.List}{EITHER A LIST CONTAINING ALL RESULTS FOR LEVELING UP FOR ***STATS*** WHICH HAVE BEEN PREGENERATED OR -1}

\item{level_Up.Repeated.Result.List}{= -1 EITHER A LIST CONTAINING ALL RESULTS FOR LEVELING UP FOR ***SKILLS*** WHICH HAVE BEEN PREGENERATED OR -1}

\item{img.Size}{STRING CONTAINING THE SIZE - "BIG" "MED" - OF THE IMAGE DESIRED.}

\item{EXAMPLE}{BOOLEAN VARIABLE. IF IT'S AN EXAMPLE, IT RETURNS -1 SO THAT THE PACKAGE CAN BE CHECKED.}
}
\description{
FUNCTION WHICH CREATES THE CHARACTER AND THEN GENERATES A CHARACTER SHEET
}
\examples{
character_CREATE( EXAMPLE = TRUE)
}
\concept{
output pdf img png character input
}
