% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_creator_trunc.R
\name{character_creator_trunc}
\alias{character_creator_trunc}
\title{A FORMAT FUNCTION}
\usage{
character_creator_trunc(x = -1, ..., prec = 0)
}
\arguments{
\item{x}{VALUE TO TRUNCATE}

\item{...}{OTHER PARAMETERS TO PASS TO base::trunc}

\item{prec}{WHAT PRECISION TO TRUNCTATE x TO}
}
\description{
FUNCTION WHICH TRUNCATES VALUES TO WHERE I WANT THEM TO BE
}
\examples{
character_creator_trunc()
}
\concept{
FORMATTING
}
