% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_LEVEL.R
\name{character_LEVEL}
\alias{character_LEVEL}
\title{A CHARACTER SHEET GENERATOR FUNCTION}
\usage{
character_LEVEL(char.Name = -1)
}
\arguments{
\item{char.Name}{NAME OF THE CHARACTER  TO LEVEL UP}
}
\description{
FUNCTION WHICH LEVELS AN ALREADY CREATED CHARACTER AND THEN GENERATES A CHARACTER SHEET
}
\examples{
character_LEVEL()
}
\concept{
level character
}
