% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_Sheet_Create.R
\name{character_Sheet_Create}
\alias{character_Sheet_Create}
\title{A PDF GENERATOR FUNCTION}
\usage{
character_Sheet_Create(species_Description_DF = -1, char.Species = -1,
  stats.End = -1, char.Realm = -1, char.Class = -1,
  char.Class.FullName = -1, char.Level = -1, char.Height = -1,
  char.Weight = -1, char.Move_Mod = -1, char.Name = -1,
  char.Player = -1, resist.Realm = -1, adept.Realm = -1,
  background_Options.Benefits = -1, skills.Final.Level = -1,
  class_Costs_And_Info = -1, char_Profession_Bonus = -1,
  skills.Primary.Stats = -1, skills.Secondary.Stats = -1,
  tp.Skills.Other_Gains = -1, tp.Name = -1, background_Options = -1,
  char.Flaws = -1, character.Life = -1, char.Age.List = -1,
  img.Size = "MED", deleteImages = TRUE)
}
\arguments{
\item{species_Description_DF}{DATAFRAME OF POSSIBLE SPECIES THE CHARACTER CAN BE}

\item{char.Species}{SPECIES THE CHARACTER IS (-1 MEANS YET TO BE DECIDED)}

\item{stats.End}{FINAL STATS OF THE CHARACTER}

\item{char.Realm}{CHARACTER'S SPELLCASTING REALM}

\item{char.Class}{CHARACTER'S CLASS}

\item{char.Class.FullName}{CHARACTER'S CLASS'S FULL NAME}

\item{char.Level}{LEVEL OF THE CHARACTER (-1 MEANS YET TO BE DECIDED)}

\item{char.Height}{NUMERIC CHARACTER'S HEIGHT}

\item{char.Weight}{NUMERIC CHARACTER'S WEIGHT}

\item{char.Move_Mod}{NUMERIC CHARACTER'S Move Mod}

\item{char.Name}{NAME OF THE CHARACTER (-1 MEANS YET TO BE DECIDED)}

\item{char.Player}{NAME OF THE PLAYER (-1 MEANS YET TO BE DECIDED)}

\item{resist.Realm}{SPELLCASTING REALM CHARACTER IS RESISTANT TO}

\item{adept.Realm}{SPELLCASTING REALM CHARACTER IS ADEPT WITH}

\item{background_Options.Benefits}{LIST CONTAINING ALL BACKGROUND OPTION BENEFITS A PLAYER HAS TAKEN}

\item{skills.Final.Level}{DATAFRAME CONTAINING ALL SKILL OUTPUTS GENERATED DURING LEVELING.}

\item{class_Costs_And_Info}{DATAFRAME CONTAINING THE COSTS AND STATS ASSOCIATED WITH SKILLS FOR ALL PROFESSIONS.}

\item{char_Profession_Bonus}{VECTOR WITH THE BONUSES TO SKILLS FROM PROFESSIONS.}

\item{skills.Primary.Stats}{DF OF THE PRIMARY SKILLS AND THE STATS ASSOCIATED WITH THEM.}

\item{skills.Secondary.Stats}{DF OF THE SECONDARY SKILLS AND THE STATS ASSOCIATED WITH THEM.}

\item{tp.Skills.Other_Gains}{TRAINING PACKAGE GAINS THAT AREN'T STATS/SKILLS}

\item{tp.Name}{TRAINING PACKAGE NAME}

\item{background_Options}{DATAFRAME OF BACKGROUND OPTIONS AVAILABLE}

\item{char.Flaws}{DATAFRAME CONTAINING ALL FLAWS THE CHARACTER TOOK}

\item{character.Life}{LIST CONTAINING THE VARIOUS FACTS OF THE CHARACTER'S LIFE}

\item{char.Age.List}{LIST CONTAINING THE CHARACTER'S AGE AT EACH LEVEL}

\item{img.Size}{STRING CONTAINING THE SIZE - "BIG" "MED" - OF THE IMAGE DESIRED.}

\item{deleteImages}{TRUE/FALSE - SHOULD THE IMAGES BE DELETED AFTER THEY HAVE CREATED THE pdf OF THE CHARACTER SHEET}
}
\description{
CREATED IMAGES OF EACH PAGE OF THE CHARACTER SHEET, OUTPUTS THEM AS A PDF, THEN DELETES THEM BY DEFAULT
}
\examples{
character_Sheet_Create()
}
\concept{
output pdf img png character
}
