% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/character_Sheet_Generator.R
\name{character_Sheet_Generator}
\alias{character_Sheet_Generator}
\title{A PDF GENERATOR FUNCTION}
\usage{
character_Sheet_Generator(char.Name = -1, deleteImages = FALSE)
}
\arguments{
\item{char.Name}{STRING CONTAINING THE CHARACTER'S NAME}

\item{deleteImages}{TRUE/FALSE SHOULD WE DELETE THE IMAGES ONCE WE'RE DONE WITH THEM.}
}
\description{
TAKES IN THE CREATED PAGES AND OUTPUTS THEM AS A PDF
}
\examples{
character_Sheet_Generator()
}
\concept{
output pdf
}
