% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_Directed_Spells_Available.R
\name{check_Directed_Spells_Available}
\alias{check_Directed_Spells_Available}
\title{A TEST FUNCTION}
\usage{
check_Directed_Spells_Available(skill.Name = -1,
  skills.Increase.Record.DF = -1)
}
\arguments{
\item{skill.Name}{NAME OF THE SKILL THEY'RE TRYING TO INCREASE}

\item{skills.Increase.Record.DF}{DATAFRAME OF SKILLS TO RECORD ANY INCREASES}
}
\description{
CHECKS TO SEE IF YOU'RE ABLE TO TAKE A PARTICULAR DIRECTED SPELL SKILL YET
}
\examples{
check_Directed_Spells_Available()
}
\concept{
input generic
}
