% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_Generic.R
\name{choose_Generic}
\alias{choose_Generic}
\title{A CHOICE FUNCTION}
\usage{
choose_Generic(choice.Name = -1, choices = -1, additional.Info = "")
}
\arguments{
\item{choice.Name}{THE NAME OF THE CHARACTER SHEET ASPECT THE USER IS CHOOSING.}

\item{choices}{A VECTOR OF THE CHOICES AVAILABLE TO THEM}

\item{additional.Info}{ANY OTHER INFO FOR THE PLAYER TO SEE}
}
\description{
GENERIC FUNCTION WHICH ASKS USER TO CHOOSE FROM A SPECIFIED LIST.
}
\examples{
choose_Generic()
}
\concept{
input generic
}
