% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_Skill_From_DF.R
\name{choose_Skill_From_DF}
\alias{choose_Skill_From_DF}
\title{A SKILL FUNCTION}
\usage{
choose_Skill_From_DF(skills.DF = -1, background_Options.Benefits = -1)
}
\arguments{
\item{skills.DF}{DATA FRAME OF THE SKILLS, SKILL COSTS, NUMBER OF TIMES YOU CAN INCREASE, NUMBER OF TIMES YOU HAVE INCREASED}

\item{background_Options.Benefits}{LIST CONTAINING ALL BACKGROUND OPTION BENEFITS A PLAYER HAS TAKEN}
}
\description{
FUNCTION TO LET THE USER CHOOSE WHICH SKILL THEY WANT TO INCREASE WITH BACKGROUND POINTS
}
\examples{
choose_Skill_From_DF()
}
\concept{
skills level
}
