% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_Skills_To_Increase_Each_Level.R
\name{choose_Skills_To_Increase_Each_Level}
\alias{choose_Skills_To_Increase_Each_Level}
\title{A SKILL FUNCTION}
\usage{
choose_Skills_To_Increase_Each_Level(skills.Increase.Record.DF = -1,
  dev_Points.Total = -1, char.Class = -1,
  background_Options.Benefits = -1, skills_Groups = -1)
}
\arguments{
\item{skills.Increase.Record.DF}{A DATAFRAME CONTAINING WHICH SKILLS HAVE BEEN INCREASED AND HOW OFTEN.}

\item{dev_Points.Total}{AN INTEGER WHICH IS HOW MANY DEV POINTS THEY HAVE THAT LEVEL}

\item{char.Class}{A STRING WHICH IS WHAT CLASS THEY ARE (DETERMINES SKILL FACTS)}

\item{background_Options.Benefits}{LIST CONTAINING ALL BACKGROUND OPTION BENEFITS A PLAYER HAS TAKEN}

\item{skills_Groups}{GROUPS OF SKILLS THAT SHARE HALF RANKS}
}
\description{
FUNCTION TO CHOOSE WHICH SKILLS TO INCREASE EACH LEVEL UNTIL YOU HAVE <10 DP REMAINING (TO GIVE A BUFFER)
}
\examples{
choose_Skills_To_Increase_Each_Level()
}
\concept{
skills level
}
