% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_Species.R
\name{choose_Species}
\alias{choose_Species}
\title{A SPECIES FUNCTION}
\usage{
choose_Species(species_Description_DF = -1)
}
\arguments{
\item{species_Description_DF}{A DATAFRAME WITH A FULL DESCRIPTION OF EACH SPECIES - BONUSES, AGE, etc...}
}
\description{
FUNCTION TO LET THE USER CHOOSE WHAT SPECIES THEY ARE
}
\examples{
choose_Species()
}
\concept{
species
}
