% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_Training_Package.R
\name{choose_Training_Package}
\alias{choose_Training_Package}
\title{A TRAINING PACKAGE FUNCTION}
\usage{
choose_Training_Package(training_Packages_Cost_DF = -1,
  training_Packages_Gains = -1, char.Class = -1, dev_Points.Total = -1)
}
\arguments{
\item{training_Packages_Cost_DF}{DATAFRAME CONTAINING INFO ON THE COST OF TRAINING PACKAGES BASED ON WHAT PROFESSION THE USER CHOSE}

\item{training_Packages_Gains}{DATAFRAME CONTAINING INFO ON THE GAINS FROM TRAINING PACKAGES}

\item{char.Class}{STRING CONTAINING THE USERS PROFESSION}

\item{dev_Points.Total}{NUMERIC VARIABLE CONTAINING THE DEVELOPMENT POINTS A CHARACTER HAS TO SPEND}
}
\description{
FUNCTION TO LET THE USER CHOOSE A SPECIFIC TRAINING PACKAGE.
}
\examples{
choose_Training_Package()
}
\concept{
skills
}
