% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{class_Costs_And_Info}
\alias{class_Costs_And_Info}
\title{Skill Costs}
\format{A data frame with 139 rows and 23 variables}
\usage{
class_Costs_And_Info
}
\description{
A dataset containing the cost for each skill, the stats associated with them and the number of age die associated with each profession because I'm bad at putting things where they should be.
}
\details{
\itemize{
  item Skill. Name for each skill 
  item Stat. Stats associated with each skill
  item Adventurer. Skill cost for profession Adventurer
  item Commoner. Skill cost for profession Commoner
  item Fighter. Skill cost for profession Fighter
  item Thief. Skill cost for profession Thief
  item Warrior_Monk. Skill cost for profession Warrior_Monk
  item Bard. Skill cost for profession Bard
  item Monk. Skill cost for profession Monk
  item Ranger. Skill cost for profession Ranger
  item Illusionist. Skill cost for profession Illusionist
  item Magician. Skill cost for profession Magician
  item Alchemist. Skill cost for profession Alchemist
  item Cleric. Skill cost for profession Cleric
  item Paladin. Skill cost for profession Paladin
  item Animist. Skill cost for profession Animist
  item Healer. Skill cost for profession Healer
  item Mentalist. Skill cost for profession Mentalist
  item Seer. Skill cost for profession Seer
  item Lay_Healer. Skill cost for profession Lay_Healer
  item Astrologer. Skill cost for profession Astrologer
  item Mystic. Skill cost for profession Mystic
  item Sorcerer. Skill cost for profession Sorcerer
}
}
\keyword{datasets}
