% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{classesDF}
\alias{classesDF}
\title{Professions}
\format{A data frame with 23 rows and 2 variables}
\usage{
classesDF
}
\description{
A dataset containing each profession
}
\details{
\itemize{
  item Number. Placeholder row
  item Profession. Name of each profession
}
}
\keyword{datasets}
