% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_Skill_Data_Frames.R
\name{create_Skill_Data_Frames}
\alias{create_Skill_Data_Frames}
\title{A SKILLS DATAFRAME CREATOR FUNCTION}
\usage{
create_Skill_Data_Frames(char.Class = -1, char.Class.FullName = -1,
  char.Realm.Name = -1, char.Realm = -1, background_Options.Benefits = -1)
}
\arguments{
\item{char.Class}{CHARACTER'S CLASS}

\item{char.Class.FullName}{CHARACTER'S CLASS'S FULLNAME, IF THEY ARE A MENTALIST MONK OR SOME SUCH}

\item{char.Realm.Name}{SPELLCASTING REALM CHARACTER USES}

\item{char.Realm}{SPELLCASTING STAT CHARACTER USES}

\item{background_Options.Benefits}{LIST CONTAINING ALL BACKGROUND OPTION BENEFITS A PLAYER HAS TAKEN}
}
\description{
FUNCTION TO CREATE THE DATA FRAMES FOR RECORDING SKILL INCREASES
}
\examples{
create_Skill_Data_Frames()
}
\concept{
skills
}
