% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crisis_Aging.R
\name{crisis_Aging}
\alias{crisis_Aging}
\title{AN AGING CRISIS FUNCTION}
\usage{
crisis_Aging(tmp = -1, pot = -1, num.Dice = -1, stat.Bonus = -1,
  stat.Dev = -1)
}
\arguments{
\item{tmp}{CHARACTER'S TEMPORARY STATS}

\item{pot}{CHARACTER'S POTENTIAL STATS}

\item{num.Dice}{NUMBER OF DICE TO BE USED DURING THE AGING CRISIS}

\item{stat.Bonus}{THE CURRENT BONUS FROM EACH STAT}

\item{stat.Dev}{THE CURRENT DEV POINTS FROM EACH STAT}
}
\description{
FUNCTION TO ADJUST STATS DURING AN AGING CRISIS.
}
\examples{
crisis_Aging()
}
\concept{
stat stats aging crisi
}
