% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{Graphics}
\name{cs.p1}
\alias{cs.p1}
\title{Character Sheet Page 1}
\format{An image, sshhhhh}
\usage{
cs.p1
}
\description{
RDA of a blank version of p1 of the character sheet.
}
\keyword{aplot}
