% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{Graphics}
\name{cs.p2}
\alias{cs.p2}
\title{Character Sheet Page 2}
\format{An image, sshhhhh}
\usage{
cs.p2
}
\description{
RDA of a blank version of p2 of the character sheet.
}
\keyword{aplot}
