% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enter_Name.R
\name{enter_Name}
\alias{enter_Name}
\title{AN INPUT READ AND CHECKING FUNCTION}
\usage{
enter_Name(character.Or.Player = -1)
}
\arguments{
\item{character.Or.Player}{STRING INDICATING WHETHER WE'RE ASKING FOR THE CHARACTER'S NAME OR THE PLAYER'S NAME}
}
\description{
FUNCTION WHICH GETS THE PLAYER TO ENTER A NON-NUMERIC NAME
}
\examples{
enter_Name()
}
\concept{
input str string
}
