% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_Page_Generator.R
\name{first_Page_Generator}
\alias{first_Page_Generator}
\title{A CHARACTER SHEET FUNCTION}
\usage{
first_Page_Generator(species_Description_DF = -1, char.Species = -1,
  stats.End = -1, char.Realm = -1, char.Class = -1, char.Level = -1,
  char.Height = -1, char.Weight = -1, char.Move_Mod = -1,
  char.Name = "", char.Player = "", resist.Realm = "", adept.Realm = "",
  background_Options.Benefits = -1, skills.OUTPUT = -1, img.Size = -1)
}
\arguments{
\item{species_Description_DF}{DATAFRAME CONTAINING BONUSES FOR STATS AND RRs FOR EACH SPECIES}

\item{char.Species}{STRING CONTAINING CHARACTER'S SPECIES}

\item{stats.End}{DATAFRAME CONTAINING CHARACTER'S STATS AFTER ALL LEVELING UP}

\item{char.Realm}{STRING CONTAINING CHARACTER'S REALM STAT}

\item{char.Class}{STRING CONTAINING CHARACTER'S PROFESSION}

\item{char.Level}{STRING CONTAINING CHARACTER'S LEVEL}

\item{char.Height}{NUMERIC CHARACTER'S HEIGHT}

\item{char.Weight}{NUMERIC CHARACTER'S WEIGHT}

\item{char.Move_Mod}{NUMERIC CHARACTER'S Move Mod}

\item{char.Name}{NUMERIC CHARACTER'S NAME}

\item{char.Player}{NUMERIC PLAYER'S NAME}

\item{resist.Realm}{STRING SPELLCASTING REALM CHARACTER IS RESISTANT TO}

\item{adept.Realm}{STRING SPELLCASTING REALM CHARACTER IS ADEPT WITH}

\item{background_Options.Benefits}{LIST CONTAINING ALL BACKGROUND OPTION BENEFITS A PLAYER HAS TAKEN}

\item{skills.OUTPUT}{DATAFRAME WITH ALL SKILLS AND RANK BONSUSES TO BE OUTPUT}

\item{img.Size}{STRING CONTAINING THE SIZE - "BIG" "MED" - OF THE IMAGE DESIRED.}
}
\description{
FUNCTION TO GENERATE THE FIRST PAGE FOR THE CHARACTER SHEET.
}
\examples{
first_Page_Generator()
}
\concept{
sheet output
}
