% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{flaws_Table}
\alias{flaws_Table}
\title{Flaws}
\format{A data frame with 22 rows and 3 variables}
\usage{
flaws_Table
}
\description{
A dataset describing the flaws a person can take to change their background options
}
\details{
\itemize{
  item Roll. Upper bound on rolls to be assigned each flaw
  item Flaw. Name of the flaw
  item Description. Description of the flaw
}
}
\keyword{datasets}
