% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_fixer.R
\name{format_fixer}
\alias{format_fixer}
\title{A FORMATTING FUNCTION}
\usage{
format_fixer(x = -1, y = -1, text.In = -1, cex.In = -1)
}
\arguments{
\item{x}{X POSITION OF THE TEXT}

\item{y}{Y POSITION OF THE TEXT}

\item{text.In}{ALL TEXT THAT THE CHARACTER SHEET WILL HAVE ON PAGE 3.}

\item{cex.In}{SIZE OF THE TEXT}
}
\description{
FUNCTION THAT FIXES THE FORMAT FOR LINES THAT WOULD GO PAST THE END OF THE PAGE.
}
\examples{
format_fixer()
}
\concept{
format output
}
