% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_Background.R
\name{generate_Background}
\alias{generate_Background}
\title{A CHARACTER BACKGROUND FUNCTION}
\usage{
generate_Background(background_Options.Benefits = -1,
  background.No_of_Children = -1, background.Sibling_Rank = -1,
  background.Family_Status = -1, background.Environment.Raised = -1,
  background.Clanhead.Relation = -1, background.Clan.Status = -1,
  background.Social_Class = -1, background.Occupation = -1,
  contact.Info = -1, char.Age = -1, char.Species = -1,
  char.Species.Raised = -1, char.Sex = -1, char.Height = -1,
  char.Weight = -1, char.Society = -1, char.Class = -1,
  skills.Increase.Record.DF = -1, background.Skills.Increase.Record.DF = -1,
  level_0_Skills.Increase.Record.List = -1)
}
\arguments{
\item{background_Options.Benefits}{LIST CONTAINING ALL BACKGROUND OPTION BENEFITS A PLAYER HAS TAKEN}

\item{background.No_of_Children}{DATAFRAME CONTAINING THE WAYS TO CALCULATE THE NUMBER OF CHILDREN FOR EACH SPECIES}

\item{background.Sibling_Rank}{DATAFRAME CONTAINING THE WAYS TO CALCULATE THE NUMBER CHILD THE PLAYER WAS}

\item{background.Family_Status}{DATAFRAME CONTAINING THE DIFFERENT POSSIBLE STATUSES FOR A FAMILY FOR EACH SPECIES, e.g., OFFSPRING, ADOPTED, etc.}

\item{background.Environment.Raised}{DATAFRAME CONTAINING THE DIFFERENT POSSIBLE FAMILY ENVIRONMENTS FOR EACH SPECIES, e.g., BASTARD, STILL TOGETHER, etc.}

\item{background.Clanhead.Relation}{DATAFRAME CONTAINING THE DIFFERENT POSSIBLE RELATIONSHIPS BETWEEN THE CHARACTER AND THE CLANHEAD}

\item{background.Clan.Status}{DATAFRAME CONTAINING THE DIFFERENT POSSIBLE STATUSES THE CHARACTER CAN HAVE IN THE CLAN}

\item{background.Social_Class}{DATAFRAME CONTAINING THE DIFFERENT POSSIBLE SOCIAL CLASSES OF THE FAMILY A CHARACTER IS RAISED IN}

\item{background.Occupation}{DATAFRAME CONTAINING THE DIFFERENT POSSIBLE OCCUPATIONS OF THE FAMILY A CHARACTER IS RAISED IN}

\item{contact.Info}{= -1 DATAFRAME CONTAINING THE DIFFERENT POSSIBLE OCCUPATIONS OF THE FAMILY A CHARACTER IS RAISED IN}

\item{char.Age}{= -1 CHARACTER'S CURRENT AGE}

\item{char.Species}{STRING CONTAINING CHARACTER'S BIRTH SPECIES}

\item{char.Species.Raised}{STRING CONTAINING SPECIES WITHIN WHOSE SOCIETY THE CHARACTER WAS RAISED}

\item{char.Sex}{STRING CONTAINING CHARACTER'S SEX}

\item{char.Height}{= -1 NUMERIC VARIABLE CONTAINING CHARACTER'S HEIGHT}

\item{char.Weight}{= -1 NUMERIC VARIABLE CONTAINING CHARACTER'S WEIGHT}

\item{char.Society}{STRING VARIABLE CONTAINING CHARACTER'S SOCIETY}

\item{char.Class}{STRING VARIABLE CONTAINING CHARACTER'S CLASS}

\item{skills.Increase.Record.DF}{DATAFRAME OF SKILLS YOU CAN INCREASE WHICH WILL GENERATE ONES USED TO RECORD ANY INCREASES}

\item{background.Skills.Increase.Record.DF}{DATAFRAME OF SKILLS YOU CAN INCREASE WHICH WILL RECORD ANY INCREASES FOR A CHARACTER'S BACKGROUND/LEVEL 0 STUFF}

\item{level_0_Skills.Increase.Record.List}{LIST CONTAINING ALL THE level_0 INFORMATION IF THAT HAS BEEN PREGENERATED}
}
\description{
FUNCTION TO GENERATE A CHARACTER'S BACKGROUND
}
\examples{
generate_Background()
}
\concept{
height weight background options society family
}
