% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Height_And_Weight_For_Species.R
\name{get_Height_And_Weight_For_Species}
\alias{get_Height_And_Weight_For_Species}
\title{A CHARACTER BACKGROUND FUNCTION}
\usage{
get_Height_And_Weight_For_Species(char.Species = -1, char.Sex = -1)
}
\arguments{
\item{char.Species}{THE SPECIES OF THE CHARACTER}

\item{char.Sex}{THE Sex OF THE CHARACTER}
}
\description{
FUNCTION TO GET THE HEIGHT AND WEIGHT FOR THE PLAYER'S SPECIES
}
\examples{
get_Height_And_Weight_For_Species()
}
\concept{
height weight
}
