% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_0_Skills.R
\name{level_0_Skills}
\alias{level_0_Skills}
\title{A LEVEL FUNCTION}
\usage{
level_0_Skills(skills.Increase.Record.DF = -1, char.Class = -1,
  hobby.Points = -1, background_Options.Benefits = -1,
  background.Social.Skills.Normal = -1, background.Social.Skills.ANY = -1,
  background.Occupation.Skills.Normal = -1,
  background.Occupation.Skills.ANY = -1, background.Skills.Assign = TRUE,
  skills_Groups = -1)
}
\arguments{
\item{skills.Increase.Record.DF}{A DATAFRAME CONTAINING WHICH SKILLS HAVE BEEN INCREASED AND HOW OFTEN.}

\item{char.Class}{A STRING WHICH IS WHAT CLASS THEY ARE (DETERMINES SKILL FACTS)}

\item{hobby.Points}{AN INTEGER WHICH IS HOW MANY HOBBY POINTS THEY HAVE.}

\item{background_Options.Benefits}{LIST CONTAINING ALL BACKGROUND OPTION BENEFITS A PLAYER HAS TAKEN}

\item{background.Social.Skills.Normal}{DATAFRAME OF THE SKILLS ASSIGNED AUTOMATICALLY FROM SOCIETY}

\item{background.Social.Skills.ANY}{DATAFRAME OF THE SKILLS ASSIGNED BY THE PLAYER FROM SOCIETY}

\item{background.Occupation.Skills.Normal}{DATAFRAME OF THE SKILLS ASSIGNED AUTOMATICALLY FROM PARENTS' OCCUPATION}

\item{background.Occupation.Skills.ANY}{DATAFRAME OF THE SKILLS ASSIGNED BY THE PLAYER FROM PARENTS' OCCUPATION}

\item{background.Skills.Assign}{TRUE/FALSE - HAVE THE BACKGROUND SKILLS ALREADY BEEN ASSIGNED}

\item{skills_Groups}{GROUPS OF SKILLS THAT SHARE HALF RANKS}
}
\description{
PLAYER ASSIGNS SKILLS FROM HOBBY POINTS AND DEVELOPMENT POINTS FOR LEVEL 0
}
\examples{
level_0_Skills()
}
\concept{
skills level hobby hobbies background
}
