% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/level_1_Skills.R
\name{level_1_Skills}
\alias{level_1_Skills}
\title{A LEVEL FUNCTION}
\usage{
level_1_Skills(skills.Increase.Record.DF = -1, dev_Points.Start = -1,
  char.Class = -1, background_Options.Benefits = -1,
  tp.Skills.Normal = -1, tp.Skills.ANY = -1, tp.Skills.Assign = TRUE,
  level_1_Skills.Increase.Record.List = -1, skills_Groups = -1)
}
\arguments{
\item{skills.Increase.Record.DF}{A DATAFRAME CONTAINING WHICH SKILLS HAVE BEEN INCREASED AND HOW OFTEN.}

\item{dev_Points.Start}{AN INTEGER WHICH IS HOW MANY DEV POINTS THEY HAVE THAT LEVEL}

\item{char.Class}{A STRING WHICH IS WHAT CLASS THEY ARE (DETERMINES SKILL FACTS)}

\item{background_Options.Benefits}{LIST CONTAINING ALL BACKGROUND OPTION BENEFITS A PLAYER HAS TAKEN}

\item{tp.Skills.Normal}{DATAFRAME OF THE SKILLS ASSIGNED AUTOMATICALLY FROM TRAINING PACKAGE}

\item{tp.Skills.ANY}{DATAFRAME OF THE SKILLS ASSIGNED BY THE PLAYER FROM TRAINING PACKAGE}

\item{tp.Skills.Assign}{TRUE/FALSE - HAVE THE TRAINING PACKAGE SKILLS ALREADY BEEN ASSIGNED}

\item{level_1_Skills.Increase.Record.List}{LIST WHICH MAY CONTAIN THE VARIABLES TO RETURNED IF THEY HAVE BEEN PREGENERATED}

\item{skills_Groups}{GROUPS OF SKILLS THAT SHARE HALF RANKS}
}
\description{
PLAYER ASSIGNS SKILLS FROM HOBBY POINTS AND DEVELOPMENT POINTS FOR LEVEL 1
}
\examples{
level_1_Skills()
}
\concept{
skills level training package
}
