% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levelUp_Skills_Stats.R
\name{levelUp_Skills_Stats}
\alias{levelUp_Skills_Stats}
\title{A LEVEL UP FUNCTION}
\usage{
levelUp_Skills_Stats(char.Level = -1, level_Up.Result.List = -1,
  stat.Tmp = -1, stat.Pot = -1, char.Age = -1, char.Age_Die = -1,
  char.Age.List = -1, char.Alive = TRUE, char.Class = -1,
  species.Middle_Aged = -1, species.Old = -1, species.Venerable = -1,
  age_Crisis.Middle_Aged = -1, age_Crisis.Old = -1,
  age_Crisis.Venerable = -1, skills.Increase.Record.DF = -1,
  background_Options.Benefits = -1, dev_Points.Level_1 = -1,
  level_1_Skills.Increase.Record.DF = -1,
  level_1_Skills.Increase.Record.List = -1,
  level_1_Skills.Increase.Record.DP_Spent = -1,
  level_Up.Repeated.Result.List = -1, skills.Each.Level.List = -1)
}
\arguments{
\item{char.Level}{STRING CONTAINING CHARACTER'S LEVEL}

\item{level_Up.Result.List}{EITHER A LIST CONTAINING ALL RESULTS FOR LEVELING UP FOR ***STATS*** WHICH HAVE BEEN PREGENERATED OR -1}

\item{stat.Tmp}{VECTOR OF CHARACTER'S TEMPORARY STATS}

\item{stat.Pot}{VECTOR OF CHARACTER'S POTENTIAL STATS}

\item{char.Age}{NUMERIC CONTAINING THE CHARACTER'S CURRENT AGE}

\item{char.Age_Die}{NUMERIC CONTAINING THE SIZE (d4, d6, etc.) OF THE CHARACTER'S AGE DIE, USED TO INCREASE THE AGE FOR EACH LEVEL GREATER THAN 1 THE CHARACTER STARTS AT.}

\item{char.Age.List}{LIST OF ALL AGES THE CHARACTER HAS REACHED EACH LEVEL SO FAR.}

\item{char.Alive}{= TRUE IS THE CHARACTER STILL ALIVE}

\item{char.Class}{STRING CONTAINING THE CHARACTER'S PROFESSION}

\item{species.Middle_Aged}{NUMERIC WITH THE AGE AT WHICH THE CHARACTER'S SPECIES REACHES MIDDLE AGE}

\item{species.Old}{NUMERIC WITH THE AGE AT WHICH THE CHARACTER'S SPECIES REACHES OLD AGE}

\item{species.Venerable}{NUMERIC WITH THE AGE AT WHICH THE CHARACTER'S SPECIES REACHES VENERABLE AGE}

\item{age_Crisis.Middle_Aged}{HAS THE CHARACTER REACHED MIDDLE AGE YET}

\item{age_Crisis.Old}{HAS THE CHARACTER REACHED OLD AGE YET}

\item{age_Crisis.Venerable}{HAS THE CHARACTER REACHED VENERABLE AGE YET}

\item{skills.Increase.Record.DF}{DATAFRAME OF SKILLS WHICH CAN BE INCREASED}

\item{background_Options.Benefits}{LIST CONTAINING ALL BACKGROUND BENEFITS CHOSEN OR ROLLED}

\item{dev_Points.Level_1}{DEVELPOMENT POINTS FOR LEVEL 1}

\item{level_1_Skills.Increase.Record.DF}{DATAFRAME OF SKILL INCREASES FROM LEVEL 1}

\item{level_1_Skills.Increase.Record.List}{RESULT OF LEVELING UP AT LEVEL 1}

\item{level_1_Skills.Increase.Record.DP_Spent}{DEVELOPMENT POINTS SPENT DURING LEVEL 1.}

\item{level_Up.Repeated.Result.List}{= -1 EITHER A LIST CONTAINING ALL RESULTS FOR LEVELING UP FOR ***SKILLS*** WHICH HAVE BEEN PREGENERATED OR -1}

\item{skills.Each.Level.List}{LIST CONTAINING DATAFRAMES FOR THE SKILLS LEVELED UP EACH LEVEL}
}
\description{
FUNCTION TO DEAL WITH LEVELING UP PAST LEVEL TO THE MAX LEVEL THE CHARACTER REACHES.
}
\examples{
levelUp_Skills_Stats()
}
\concept{
level skills stats
}
