% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{occupation_Unguilded}
\alias{occupation_Unguilded}
\title{Unguilded Occupations}
\format{A data frame with 22 rows and 12 variables}
\usage{
occupation_Unguilded
}
\description{
A dataset describing the unguilded occupation for each people and bonuses from them.
}
\details{
\itemize{
  item Tribal. Upper bound on rolls to be assigned a job for Tribal people 
  item Viking. Upper bound on rolls to be assigned a job for Viking people 
  item Feudal. Upper bound on rolls to be assigned a job for Feudal people 
  item Imperial. Upper bound on rolls to be assigned a job for Imperial people 
  item Elf. Upper bound on rolls to be assigned a job for Elf people 
  item Dwarf. Upper bound on rolls to be assigned a job for Dwarf people 
  item Hobbit. Upper bound on rolls to be assigned a job for Hobbit people 
  item Gnome. Upper bound on rolls to be assigned a job for Gnome people 
  item Occupation. Name of an occupation
  item Ur.. Percentage chance of the profession being urban
  item Skills. Names of skills you gain ranks in
  item Ranks. Number of skill ranks you gain for each skill
}
}
\keyword{datasets}
