% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{pace_Multiplier}
\alias{pace_Multiplier}
\title{Pace Multiplier}
\format{A data frame with 6 rows and 5 variables}
\usage{
pace_Multiplier
}
\description{
A dataset describing what rates of movement are allowed or disallowed
}
\details{
\itemize{
  item Pace. Title of each rate of movement
  item Multiplier. Multiplier of each rate of movement
  item Difficulty. Difficulty of each rate of movement
  item Armor.Qu.Penalty.Prohibited.Paces. Prohibited rates of movement for levels of quickness penalty
  item Encumbrance.Penalty.Prohibited.Paces. Prohibited rates of movement for levels of encumbrance
}
}
\keyword{datasets}
