% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{potential_Stats_Generator_DF}
\alias{potential_Stats_Generator_DF}
\title{Potential Stats Generation}
\format{A data frame with 17 rows and 4 variables}
\usage{
potential_Stats_Generator_DF
}
\description{
A dataset describing how to generate potential stats for each temporary stat range
}
\details{
\itemize{
  item Stat. Temporary stat value
  item Base_Pot. Base for the potential stat associated with each temporary stat
  item Number.of.Dice. Number of dice to roll associated with each temporary stat
  item Number.of.Sides. Number of sides for each die rolled associated with each temporary stat
}
}
\keyword{datasets}
