% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_Bonus_Calculator.R
\name{rank_Bonus_Calculator}
\alias{rank_Bonus_Calculator}
\title{A SKILL BONUS FUNCTION}
\usage{
rank_Bonus_Calculator(x = -1, bd.Progress = -1, char.Species = -1,
  char.Level = -1)
}
\arguments{
\item{x}{THE ROW CONTAINING THE RANKS INCREASED AND THE SKILL NAME AND OTHER DETAILS.}

\item{bd.Progress}{THE RATE AT WHICH BODY DEVELOPMENT INCREASES FOR THE CHARACTER'S SPECIES.}

\item{char.Species}{STRING CONTAINING CHARACTER'S SPECIES}

\item{char.Level}{NUMERIC CONTAINING CHARACTER'S LEVEL}
}
\description{
FUNCTION TO CALCULATE THE BONUSES TO A SKILL FOR A RANK.
}
\examples{
rank_Bonus_Calculator()
}
\concept{
bonus skill
}
