% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Integer.R
\name{read_Integer}
\alias{read_Integer}
\title{AN INPUT READ AND CHECKING FUNCTION}
\usage{
read_Integer(val.Min = -1, val.Max = -1)
}
\arguments{
\item{val.Min}{MINIMUM VALUE THE USER CAN ENTER}

\item{val.Max}{MAXIMUM VALUE THE USER CAN ENTER}
}
\description{
FUNCTION TO READ AN INTEGER FROM THE USER, BETWEEN TWO VALUES.
}
\examples{
read_Integer()
}
\concept{
input int integer
}
