% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_Yes_Or_No.R
\name{read_Yes_Or_No}
\alias{read_Yes_Or_No}
\title{AN INPUT READ AND CHECKING FUNCTION}
\usage{
read_Yes_Or_No(prom = -1)
}
\arguments{
\item{prom}{STRING WITH THE PROMPT TO OUTPUT TO THE SCREEN}
}
\description{
FUNCTION TO ACCEPT SOME INDICATOR FOR "YES" OR "NO" FROM A USER AND RETURN "Yes" OR "No"
}
\examples{
read_Yes_Or_No()
}
\concept{
input string
}
