% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recreate_Command_Generator.R
\name{recreate_Command_Generator}
\alias{recreate_Command_Generator}
\title{A RECREATE FUNCTION}
\usage{
recreate_Command_Generator(recreate.Command = -1, char.Name = "")
}
\arguments{
\item{recreate.Command}{LIST CONTAINING ALL THE CHARACTER'S DETAILS.}

\item{char.Name}{STRING CONTAINING THE CHARACTER'S NAME}
}
\description{
FUNCTION TO GENERATE A JSON CONTAINING ALL THE CHARACTER'S DETAILS
}
\examples{
recreate_Command_Generator()
}
\concept{
recreate
}
