% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{rm_Levels_Table}
\alias{rm_Levels_Table}
\title{Levels}
\format{A data frame with 21 rows and 2 variables}
\usage{
rm_Levels_Table
}
\description{
A dataset of experience needed for each level up
}
\details{
\itemize{
  item Level. Placeholder row
  item rm_levels. Experience needed to reach each level 
}
}
\keyword{datasets}
