% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/second_Page_Generator.R
\name{second_Page_Generator}
\alias{second_Page_Generator}
\title{A CHARACTER SHEET FUNCTION}
\usage{
second_Page_Generator(skills.OUTPUT = -1, class_Costs_And_Info = -1,
  char_Profession_Bonus = -1, char.Class = -1, skills.Primary.Stats = -1,
  skills.Secondary.Stats = -1, bonus.Total = -1, char.Realm = -1,
  char.Species = -1, species_Description_DF = -1, img.Size = -1)
}
\arguments{
\item{skills.OUTPUT}{DATAFRAME CONTAINING ALL SKILL OUTPUTS GENERATED DURING LEVELING.}

\item{class_Costs_And_Info}{DATAFRAME CONTAINING THE COSTS AND STATS ASSOCIATED WITH SKILLS FOR ALL PROFESSIONS.}

\item{char_Profession_Bonus}{VECTOR WITH THE BONUSES TO SKILLS FROM PROFESSIONS.}

\item{char.Class}{STRING CONTAINING CHARACTER'S PROFESSION.}

\item{skills.Primary.Stats}{VECTOR OF THE PRIMARY SKILLS.}

\item{skills.Secondary.Stats}{VECTOR OF THE SECONDARY SKILLS.}

\item{bonus.Total}{TOTAL BONUS ASSOCIATED WITH EACH STAT.}

\item{char.Realm}{STRING CONTAINING CHARACTER'S REALM STAT}

\item{char.Species}{STRING CONTAINING CHARACTER'S SPECIES}

\item{species_Description_DF}{DATAFRAME CONTAINING BONUSES FOR STATS AND RRs FOR EACH SPECIES}

\item{img.Size}{STRING CONTAINING THE SIZE - "BIG" "MED" - OF THE IMAGE DESIRED.}
}
\description{
FUNCTION TO GENERATE THE SECOND PAGE FOR THE CHARACTER SHEET.
}
\examples{
second_Page_Generator()
}
\concept{
test skill
}
