% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills_DF_OUT_CREATE.R
\name{skills_DF_OUT_CREATE}
\alias{skills_DF_OUT_CREATE}
\title{A SKILL DATAFRAME OUTPUT FUNCTION}
\usage{
skills_DF_OUT_CREATE(skills.Increase.Record.DF = -1, increase = TRUE)
}
\arguments{
\item{skills.Increase.Record.DF}{DATA FRAME OF THE SKILLS, SKILL COSTS, NUMBER OF TIMES YOU CAN INCREASE, NUMBER OF TIMES YOU HAVE INCREASED}

\item{increase}{BOOLEAN VALUE FOR WHETHER WE ARE INCREASING THE NUMBER OF SKILLS TAKEN THAT LEVEL OR NOT. DEFAULTS TO INCREASING}
}
\description{
FUNCTION TO CREATE THE SKILL DATAFRAME THAT GETS OUTPUT TO THE SCREEN
}
\examples{
skills_DF_OUT_CREATE()
}
\concept{
skill skills output
}
