% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{skills_GroupsDF}
\alias{skills_GroupsDF}
\title{Skill groups}
\format{A data frame with 4 rows and 2 variables}
\usage{
skills_GroupsDF
}
\description{
A dataset of skill groups
}
\details{
\itemize{
  item Number. Placeholder row
  item Groups. Skills groups which share half ranks
}
}
\keyword{datasets}
