% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{species_Description_DF}
\alias{species_Description_DF}
\title{Species Description}
\format{A data frame with 12 rows and 25 variables}
\usage{
species_Description_DF
}
\description{
A dataset for bonuses and penalties and age relations for each species
}
\details{
\itemize{
  item Race. The name of each race
  item Constitution. Co bonus for each race
  item Agility. Ag bonus for each race
  item Self_Discipline. SD bonus for each race
  item Reasoning. Re bonus for each race
  item Memory. Me bonus for each race
  item Strength. St bonus for each race
  item Quickness. Qu bonus for each race
  item Empathy. Em bonus for each race
  item Intuition. In bonus for each race
  item Presence. Pr bonus for each race
  item Essense. Realm bonus for each race
  item Channelling. Realm bonus for each race
  item Mentalism. Realm bonus for each race
  item Poison. Resistance bonus for each race
  item Disease. Resistance bonus for each race
  item Background. Background Options Points for each race
  item Body_Dev. Body development bonus for each race
  item Adolescent. Age when they reach adolescence
  item Middle_Aged. Age when they reach middle age
  item Old. Age when they reach old age
  item Venerable  Age when they reach adolescence
  item Maximum_Age. 
}
}
\keyword{datasets}
