% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{spell_Lists_Base}
\alias{spell_Lists_Base}
\title{Base Spell Lists}
\format{A data frame with 9 rows and 21 variables}
\usage{
spell_Lists_Base
}
\description{
A dataset for the base spell lists for each profession
}
\details{
\itemize{
  item Adventurer. Base spell lists for an Adventurer
  item Commoner. Base spell lists for a Commoner
  item Fighter. Base spell lists for a Fighter
  item Thief. Base spell lists for a Thief
  item Warrior_Monk. Base spell lists for a Warrior_Monk
  item Bard. Base spell lists for a Bard
  item Monk. Base spell lists for a Monk
  item Ranger. Base spell lists for a Ranger
  item Illusionist. Base spell lists for a Illusionist
  item Magician. Base spell lists for a Magician
  item Elementalist. Base spell lists for an Elementalist
  item Alchemist_Channeling. Base spell lists for an Alchemist
  item Alchemist. Base spell lists for an Alchemist
  item Cleric. Base spell lists for a Cleric
  item Paladin. Base spell lists for a Paladin
  item Animist. Base spell lists for an Animist
  item Healer. Base spell lists for a Healer
  item Mentalist. Base spell lists for a Mentalist
  item Seer. Base spell lists for a Seer
  item Lay_Healer. Base spell lists for a Lay_Healer
  item Astrologer. Base spell lists for an Astrologer
  item Mystic. Base spell lists for a Mystic
  item Sorcerer. Base spell lists for a Sorcerer 
}
}
\keyword{datasets}
