% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{spell_Lists_Closed}
\alias{spell_Lists_Closed}
\title{Closed Spell Lists}
\format{A data frame with 10 rows and 3 variables}
\usage{
spell_Lists_Closed
}
\description{
A dataset for the closed spell lists in each realm
}
\details{
\itemize{
  item Channeling. Name of each closed Channeling Spell list
  item Essence. Name of each closed Essence Spell list
  item Mentalism. Name of each closed Mentalism Spell list
}
}
\keyword{datasets}
