% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{spell_Lists_Evil}
\alias{spell_Lists_Evil}
\title{Evil Spell Lists}
\format{A data frame with 6 rows and 3 variables}
\usage{
spell_Lists_Evil
}
\description{
A dataset for the evil spell lists in each realm
}
\details{
\itemize{
  item Channeling. Name of each evil Channeling Spell list
  item Essence. Name of each evil Essence Spell list
  item Mentalism. Name of each evil Mentalism Spell list
}
}
\keyword{datasets}
