% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{spell_Lists_Open}
\alias{spell_Lists_Open}
\title{Open Spell Lists}
\format{A data frame with 10 rows and 3 variables}
\usage{
spell_Lists_Open
}
\description{
A dataset for the open spell lists in each realm
}
\details{
\itemize{
  item Channeling. Name of each open Channeling Spell list
  item Essence. Name of each open Essence Spell list
  item Mentalism. Name of each open Mentalism Spell list
}
}
\keyword{datasets}
