% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{spell_Lists_Shock_Bolt}
\alias{spell_Lists_Shock_Bolt}
\title{Spell lists for Shock Bolt}
\format{A data frame with 5 rows and 2 variables}
\usage{
spell_Lists_Shock_Bolt
}
\description{
A dataset for the spell lists you must know up to level n to learn directed shock spells
}
\details{
\itemize{
  item Spell.List. Name of each spell list
  item Level. Level you must have in it to learn directed spells
}
}
\keyword{datasets}
