% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{stats_Primary}
\alias{stats_Primary}
\title{Primary Stats Data}
\format{A data frame with 2 rows and 22 variables}
\usage{
stats_Primary
}
\description{
A dataset containing primary stats for each profession
}
\details{
\itemize{
  item Adventurer. Primary stats for an Adventurer
  item Commoner. Primary stats for a Commoner
  item Fighter. Primary stats for a Fighter
  item Thief. Primary stats for a Thief
  item Warrior_Monk. Primary stats for a Warrior_Monk
  item Bard. Primary stats for a Bard
  item Monk. Primary stats for a Monk
  item Ranger. Primary stats for a Ranger
  item Illusionist. Primary stats for a Illusionist
  item Magician. Primary stats for a Magician
  item Alchemist. Primary stats for an Alchemist
  item Cleric. Primary stats for a Cleric
  item Paladin. Primary stats for a Paladin
  item Animist. Primary stats for an Animist
  item Healer. Primary stats for a Healer
  item Mentalist. Primary stats for a Mentalist
  item Seer. Primary stats for a Seer
  item Lay_Healer. Primary stats for a Lay_Healer
  item Astrologer. Primary stats for an Astrologer
  item Mystic. Primary stats for a Mystic
  item Sorcerer. Primary stats for a Sorcerer 
}
}
\keyword{datasets}
