% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolemasterCharacterCreator.R
\docType{data}
\name{statVal_And_Bonus_And_DP}
\alias{statVal_And_Bonus_And_DP}
\title{Stat Ranges Data}
\format{A data frame with 32 rows and 3 variables}
\usage{
statVal_And_Bonus_And_DP
}
\description{
A dataset ranges of stat values and the bonuses and DP associated with them.
}
\details{
\itemize{
  item Stat_Values. Ranges of stats associated with different bonuses and development points
  item Bonus. Bonuses associated with different ranges of stats 
  item DP. Development points associated with different ranges of stats 
}
}
\keyword{datasets}
